Update README.md to run docker correctly
Original readme is wrong with change port command. Internal change is to 4001, but this new port is not used in the port mapping parameter.

Secondly, the container is hard to stop. ctrl-c doesn't work because `--init` wasn't used. And `docker stop` is hard to use because container didn't have a name.